worker/requirements.txt
httpx>=0.27.0
supabase>=2.10.0
python-dotenv>=1.0.0
yt-dlp>=2024.1.0
Pillow>=10.0.0
structlog>=24.1.0
google-generativeai>=0.8.0
anthropic>=0.40.0
